Rebuild of the broken merge 1123aceba, which was committed with 193
unresolved conflict files. All conflicts resolved in favour of
version-3.2 (-X theirs); prod-only docs and nginx configs are kept.

+++ cmdb/interface/rest_api/routes/media_library_routes/media_file_routes.py
@@ -1,5 +1,5 @@
-# DATAGERRY - OpenSource Enterprise CMDB
-# Copyright (C) 2025 becon GmbH
+# DataGerry - OpenSource Enterprise CMDB
+# Copyright (C) 2026 becon GmbH
 #
 # This program is free software: you can redistribute it and/or modify
 # it under the terms of the GNU Affero General Public License as
@@ -17,10 +17,13 @@
 Implementation of all API routes for the MediaFiles
 """
 import json
-import logging
+from logging import Logger, getLogger
 from bson import json_util
 from flask import abort, request, Response
+from werkzeug.wrappers.response import Response as Resp
+from werkzeug.exceptions import HTTPException
 
+from cmdb.interface.rest_api.responses.gridfs_response import GridFsResponse
 from cmdb.manager.manager_provider_model import ManagerProvider, ManagerType
 from cmdb.manager import MediaFilesManager
 
@@ -55,18 +58,18 @@ from cmdb.errors.manager.media_files_manager import (
 )
 # -------------------------------------------------------------------------------------------------------------------- #
 
-LOGGER = logging.getLogger(__name__)
+LOGGER: Logger = getLogger(__name__)
 
 media_file_blueprint = APIBlueprint('media_file_blueprint', __name__, url_prefix='/media_file')
 
 # -------------------------------------------------------------------------------------------------------------------- #
 
 @media_file_blueprint.route('/', methods=['GET', 'HEAD'])
-@media_file_blueprint.parse_collection_parameters()
 @insert_request_user
-@media_file_blueprint.protect(auth=True, right='base.framework.object.view')
 @verify_api_access(required_api_level=ApiLevel.LOCKED)
-def get_file_list(params: CollectionParameters, request_user: CmdbUser):
+@media_file_blueprint.protect(auth=True, right='base.framework.object.view')
+@media_file_blueprint.parse_collection_parameters()
+def get_file_list(params: CollectionParameters, request_user: CmdbUser) -> Resp:
     """
     Get all objects in database
 
@@ -82,11 +85,13 @@ def get_file_list(params: CollectionParameters, request_user: CmdbUser):
 
         metadata = generate_collection_parameters(params=params)
         response_query = {'limit': params.limit, 'skip': params.skip, 'sort': [(params.sort, params.order)]}
-        output = media_files_manager.get_many_media_files(metadata, **response_query)
+        output: GridFsResponse = media_files_manager.get_many_media_files(metadata, **response_query)
 
         api_response = GetMultiResponse(output.result, total=output.total, params=params, url=request.url)
 
         return api_response.make_response()
+    except HTTPException as http_err:
+        raise http_err
     except MediaFileManagerGetError as err:
         LOGGER.error("[get_file_list] MediaFileManagerGetError: %s", err, exc_info=True)
         abort(400, "Failed to retrieve the FilesList from the database!")
@@ -99,7 +104,7 @@ def get_file_list(params: CollectionParameters, request_user: CmdbUser):
 @insert_request_user
 @verify_api_access(required_api_level=ApiLevel.LOCKED)
 @right_required('base.framework.object.edit')
-def add_new_file(request_user: CmdbUser):
+def add_new_file(request_user: CmdbUser) -> Resp:
     """
     This method saves a file to the specified section of the document for storing workflow data.
     Any existing value that matches filename and the metadata is deleted. Before saving a value.
@@ -146,7 +151,8 @@ def add_new_file(request_user: CmdbUser):
 
         if file_exists:
             exist = media_files_manager.get_file(filter_metadata)
-            media_files_manager.delete_file(exist['public_id'])
+            if exist:
+                media_files_manager.delete_file(exist['public_id'])
 
         # If file exist overwrite the references from previous file
         if exist:
@@ -159,6 +165,8 @@ def add_new_file(request_user: CmdbUser):
         result = media_files_manager.insert_file(data=file, metadata=metadata)
 
         return InsertSingleResponse(result, result['public_id']).make_response()
+    except HTTPException as http_err:
+        raise http_err
     except MediaFileManagerGetError as err:
         LOGGER.error("[add_new_file] MediaFileManagerGetError: %s", err, exc_info=True)
         abort(400, "Failed to retrieve the FilesList from the database!")
@@ -174,7 +182,7 @@ def add_new_file(request_user: CmdbUser):
 @insert_request_user
 @verify_api_access(required_api_level=ApiLevel.LOCKED)
 @right_required('base.framework.object.edit')
-def update_file(request_user: CmdbUser):
+def update_file(request_user: CmdbUser) -> Resp:
     """
     This method updates a file to the specified section in the document.
     Any existing value that matches the file name and metadata is taken into account.
@@ -229,6 +237,8 @@ def update_file(request_user: CmdbUser):
         media_files_manager.update_file(data)
 
         return DefaultResponse(data).make_response()
+    except HTTPException as http_err:
+        raise http_err
     except MediaFileManagerUpdateError as err:
         LOGGER.error("[update_file] MediaFileManagerUpdateError: %s", err, exc_info=True)
         abort(400, "Failed to update the File in the database!")
@@ -242,7 +252,7 @@ def update_file(request_user: CmdbUser):
 @insert_request_user
 @verify_api_access(required_api_level=ApiLevel.LOCKED)
 @media_file_blueprint.protect(auth=True, right='base.framework.object.view')
-def get_file(filename: str, request_user: CmdbUser):
+def get_file(filename: str, request_user: CmdbUser) -> Resp:
     """
     This method fetch a file to the specified section of the document.
     Any existing value that matches the file name and metadata will be considered.
@@ -268,6 +278,8 @@ def get_file(filename: str, request_user: CmdbUser):
             result = None
 
         return DefaultResponse(result).make_response()
+    except HTTPException as http_err:
+        raise http_err
     except Exception as err:
         LOGGER.error("[get_file] Exception: %s. Type: %s", err, type(err), exc_info=True)
         abort(500, f"An internal server error occured while retrieving the file: {filename}!")
@@ -277,7 +289,7 @@ def get_file(filename: str, request_user: CmdbUser):
 @insert_request_user
 @verify_api_access(required_api_level=ApiLevel.LOCKED)
 @media_file_blueprint.protect(auth=True, right='base.framework.object.view')
-def download_file(filename: str, request_user: CmdbUser):
+def download_file(filename: str, request_user: CmdbUser) -> Resp:
     """
     This method download a file to the specified section of the document.
     Any existing value that matches the file name and metadata will be considered.
@@ -306,16 +318,19 @@ def download_file(filename: str, request_user: CmdbUser):
                     f"attachment; filename={filename}"
             }
         )
+    except HTTPException as http_err:
+        raise http_err
     except Exception as err:
         LOGGER.error("[download_file] Exception: %s. Type: %s", err, type(err), exc_info=True)
         abort(500, f"An internal server error occured while downloading the file: {filename}!")
 
 # --------------------------------------------------- CRUD - DELETE -------------------------------------------------- #
 
-@media_file_blueprint.route('<int:public_id>', methods=['DELETE'])
+@media_file_blueprint.route('/<int:public_id>', methods=['DELETE'])
 @insert_request_user
+@verify_api_access(required_api_level=ApiLevel.LOCKED)
 @media_file_blueprint.protect(auth=True, right='base.framework.object.edit')
-def delete_file(public_id: int, request_user: CmdbUser):
+def delete_file(public_id: int, request_user: CmdbUser) -> Resp:
     """
     This method deletes a file in the specified section of the document for storing workflow data.
     Any existing value that matches the file name and metadata is deleted. Before saving a value.
@@ -341,6 +356,8 @@ def delete_file(public_id: int, request_user: CmdbUser):
                 media_files_manager.delete_file(_id)
 
         return DefaultResponse(file_to_delete).make_response()
+    except HTTPException as http_err:
+        raise http_err
     except MediaFileManagerDeleteError as err:
         LOGGER.error("[delete_file] MediaFileManagerDeleteError: %s", err, exc_info=True)
         abort(400, f"Failed to delete the File with ID: {public_id} in the database!")
